Move homepage components into a dedicated home directory
