+++ src/lib/components/home/hero.svelte
